Add bulk of the reference text and have most of it generated
