+++ book/src/SUMMARY.md
@@ -11,15 +11,15 @@
     - [Compilation](./v2/reference-compilation.md)
     - [Language]()
       - [Tokens & AST](./v2/reference-language-tokens_ast.md)
-      - [Manifest]()
-      - [Device]()
-      - [Block]()
-      - [Register]()
-      - [Command]()
-      - [Buffer]()
-      - [Fieldset]()
-      - [Enum]()
-      - [Extern]()
+      - [Manifest](./v2/reference-language-manifest.md)
+      - [Device](./v2/reference-language-device.md)
+      - [Block](./v2/reference-language-block.md)
+      - [Register](./v2/reference-language-register.md)
+      - [Command](./v2/reference-language-command.md)
+      - [Buffer](./v2/reference-language-buffer.md)
+      - [Fieldset](./v2/reference-language-fieldset.md)
+      - [Enum](./v2/reference-language-enum.md)
+      - [Extern](./v2/reference-language-extern.md)
     - [Runtime]()
       - [Rust]()
