+++ book/src/gen-docs/mir-shapes/block.md
@@ -1,0 +1,31 @@
+## Shape
+> todo: example
+
+| Property | Value |
+| --- | --- |
+| Identifier namespace | `All` |
+| Supports repeat | `yes` |
+| Supports basetype | `no` |
+| Supports conversion type | `no` |
+| Supports short properties | `no` |
+| Supports properties | `yes`, see below |
+| Supports subnodes | `yes`, see below |
+## Long properties
+These properties are specified in the node body.
+### address-offset
+> todo: description
+
+- required: `yes`
+- multiple allowed: `no`
+- supports doc comments: `no`
+#### Allowed expression types
+- number: `0`
+## Possible subnodes
+Subnodes of the following types are allowed in the node body.
+- [block]
+- [register]
+- [command]
+- [buffer]
+- [fieldset]
+- [enum]
+- [extern]
