+++ book/src/gen-docs/mir-shapes/buffer.md
@@ -1,0 +1,30 @@
+## Shape
+> todo: example
+
+| Property | Value |
+| --- | --- |
+| Identifier namespace | `Operation` |
+| Supports repeat | `no` |
+| Supports basetype | `no` |
+| Supports conversion type | `no` |
+| Supports short properties | `no` |
+| Supports properties | `yes`, see below |
+| Supports subnodes | `no` |
+## Long properties
+These properties are specified in the node body.
+### access
+> todo: description
+
+- required: `no`
+- multiple allowed: `no`
+- supports doc comments: `no`
+#### Allowed expression types
+- access specifier: `RW`
+### address
+> todo: description
+
+- required: `yes`
+- multiple allowed: `no`
+- supports doc comments: `no`
+#### Allowed expression types
+- number: `0`
