+++ book/src/gen-docs/mir-shapes/enum.md
@@ -1,0 +1,27 @@
+## Shape
+> todo: example
+
+| Property | Value |
+| --- | --- |
+| Identifier namespace | `Type` |
+| Supports repeat | `no` |
+| Supports basetype | `yes` |
+| Supports conversion type | `no` |
+| Supports short properties | `no` |
+| Supports properties | `yes`, see below |
+| Supports subnodes | `no` |
+## Long properties
+These properties are specified in the node body.
+### *any name*
+> todo: description
+
+- required: `no`
+- multiple allowed: `yes`
+- supports doc comments: `yes`
+#### Allowed expression types
+- auto: `_`
+- number: `0`
+- default number: `default 0`
+- default auto: `default _`
+- catch-all number: `catch-all 0`
+- catch-all auto: `catch-all _`
