+++ book/src/gen-docs/mir-shapes/extern.md
@@ -1,0 +1,30 @@
+## Shape
+> todo: example
+
+| Property | Value |
+| --- | --- |
+| Identifier namespace | `Type` |
+| Supports repeat | `no` |
+| Supports basetype | `yes` |
+| Supports conversion type | `no` |
+| Supports short properties | `no` |
+| Supports properties | `yes`, see below |
+| Supports subnodes | `no` |
+## Long properties
+These properties are specified in the node body.
+### infallible
+> todo: description
+
+- required: `no`
+- multiple allowed: `no`
+- supports doc comments: `no`
+#### Allowed expression types
+- allow: `allow`
+### size-bits
+> todo: description
+
+- required: `no`
+- multiple allowed: `no`
+- supports doc comments: `no`
+#### Allowed expression types
+- number: `8`
