+++ book/src/gen-docs/mir-shapes/field.md
@@ -1,0 +1,31 @@
+## Shape
+> todo: example
+
+| Property | Value |
+| --- | --- |
+| Identifier namespace | `All` |
+| Supports repeat | `yes` |
+| Supports basetype | `yes` |
+| Supports conversion type | `yes` |
+| Supports short properties | `yes`, see below |
+| Supports properties | `no` |
+| Supports subnodes | `no` |
+## Short properties
+These properties are specified inline in the node definition and are used without name.
+### address
+> todo: description
+
+- required: `yes`
+- multiple allowed: `no`
+- supports doc comments: `no`
+#### Allowed expression types
+- number: `0`
+- range: `8:0`
+### access
+> todo: description
+
+- required: `no`
+- multiple allowed: `no`
+- supports doc comments: `no`
+#### Allowed expression types
+- access specifier: `RW`
