+++ compiler/dd-core/Cargo.toml
@@ -33,4 +33,4 @@ pretty_assertions = "1.4.0"
 [features]
 default = ["prettyplease"]
 prettyplease = ["dep:prettyplease", "dep:syn"]
-gen-docs = ["device-driver-parser/gen-docs"]
+gen-docs = ["device-driver-parser/gen-docs", "device-driver-mir/gen-docs"]
