+++ compiler/dd-mir/Cargo.toml
@@ -17,3 +17,7 @@ convert_case.workspace = true
 itertools.workspace = true
 fastrand.workspace = true
 clap.workspace = true
+
+[features]
+default = ["gen-docs"]
+gen-docs = []
