+++ compiler/dd-mir/src/lib.rs
@@ -15,6 +15,9 @@ mod lowering;
 pub mod model;
 pub(crate) mod passes;
 
+#[cfg(feature = "gen-docs")]
+pub use lowering::gen_docs::gen_docs;
+
 #[derive(Parser, Debug, Clone, Default)]
 #[command(no_binary_name = true)]
 pub struct MirOptions {
