+++ compiler/dd-parser/src/lib.rs
@@ -261,14 +261,16 @@ impl<'src> Display for Expression<'src> {
             Expression::Integer(_) => write!(f, "integer type"),
             Expression::Allow => write!(f, "allow"),
             Expression::Number(_) => write!(f, "number"),
-            Expression::DefaultNumber(_) => write!(f, "default number"),
-            Expression::CatchAllNumber(_) => write!(f, "catch-all number"),
+            Expression::DefaultNumber(None) => write!(f, "default auto"),
+            Expression::CatchAllNumber(None) => write!(f, "catch-all auto"),
+            Expression::DefaultNumber(Some(_)) => write!(f, "default number"),
+            Expression::CatchAllNumber(Some(_)) => write!(f, "catch-all number"),
             Expression::String(_) => write!(f, "string"),
             Expression::Access(_) => write!(f, "access specifier"),
             Expression::ByteOrder(_) => write!(f, "byte order"),
             Expression::TypeReference(_) => write!(f, "type reference"),
             Expression::SubNode(_) => write!(f, "sub node"),
-            Expression::Auto => write!(f, "_"),
+            Expression::Auto => write!(f, "auto"),
             Expression::AddressMode(_) => write!(f, "address mode"),
             Expression::Error => write!(f, "error"),
         }
